Build fix

+++ include/aws/s3/private/s3_client_impl.h
@@ -91,7 +91,7 @@ struct aws_s3_client {
     struct aws_tls_connection_options *tls_connection_options;
 
     /* Timeout value, in milliseconds, used for each connection. */
-    const uint64_t connection_timeout_ms;
+    const uint32_t connection_timeout_ms;
 
     /* Throughput target in Gbps that we are trying to reach. */
     const double throughput_target_gbps;
